style: reformat jinadRuntime
