+++ .dockerignore
@@ -8,14 +8,5 @@
 !*requirements.txt
 !README.md
 !MANIFEST.in
-!tests/**
 !cli/**
-
-# Ignore unnecessary files inside allowed directories
-# This should go after the allowed directories
-
-**/*.log
-**/Thumbs.db
-**/.pyre/
-**/.eggs/
-__pycache__
+!daemon/**
