+++ Dockerfiles/alpinex.Dockerfile
@@ -16,19 +16,18 @@ LABEL org.opencontainers.image.created=$BUILD_DATE \
       org.opencontainers.image.title="Jina" \
       org.opencontainers.image.description="Jina is the cloud-native neural search solution powered by state-of-the-art AI and deep learning technology"
 
-WORKDIR /jina/
-
-ADD setup.py MANIFEST.in requirements.txt extra-requirements.txt README.md ./
-ADD cli ./cli/
-ADD jina ./jina/
+COPY . /jina/
 
 ENV PYTHONPATH=$PYTHONPATH:/usr/lib/python3.8/dist-packages:/usr/local/lib/python3.8/site-packages:/usr/lib/python3/dist-packages:/usr/local/lib/python3/site-packages \
     PIP_NO_CACHE_DIR=1 \
-    PIP_DISABLE_PIP_VERSION_CHECK=1
+    PIP_DISABLE_PIP_VERSION_CHECK=1 \
+    JINA_VCS_VERSION=$VCS_REF \
+    JINA_BUILD_DATE=$BUILD_DATE
 
 # py3-scipy
 RUN apk add --no-cache py3-pyzmq py3-numpy grpc && \
     ln -s locale.h /usr/include/xlocale.h && \
+    cd /jina && \
     pip install . --compile && \
     find /usr/lib/python3.8/ -name 'tests' -exec rm -r '{}' + && \
     find /usr/lib/python3.8/site-packages/ -name '*.so' -print -exec sh -c 'file "{}" | grep -q "not stripped" && strip -s "{}"' \; && \
@@ -36,10 +35,5 @@ RUN apk add --no-cache py3-pyzmq py3-numpy grpc && \
     rm -rf /tmp/* && \
     rm -rf /jina && \
     rm -rf /var/cache/apk/*
-
-WORKDIR /
-
-ENV JINA_VCS_VERSION=$VCS_REF \
-    JINA_BUILD_DATE=$BUILD_DATE
 
 ENTRYPOINT ["jina"]
