+++ Dockerfiles/pip.Dockerfile
@@ -1,22 +1,14 @@
 ARG PY_VERSION=3.7
+ARG PIP_TAG
 
 FROM python:${PY_VERSION}-slim
 
 RUN apt-get update && apt-get install --no-install-recommends -y gcc libc6-dev
 
-WORKDIR /jina/
+COPY . /jina/
 
-ADD setup.py MANIFEST.in requirements.txt extra-requirements.txt README.md ./
-ADD cli ./cli/
-ADD jina ./jina/
-
-ARG PIP_TAG
-
-RUN pip install ."$PIP_TAG"
-
+RUN cd /jina && pip install ."$PIP_TAG"
 RUN cat $HOME/.bashrc
 RUN grep -Fxq "# JINA_CLI_BEGIN" $HOME/.bashrc
-
-WORKDIR /
 
 ENTRYPOINT ["jina"]
